requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
torch>=2.0.0
hmmlearn>=0.3.0
arch>=6.0.0
ta>=0.10.2
websockets>=11.0.3
aiohttp>=3.8.5
python-telegram-bot>=20.0
pyyaml>=6.0
python-dotenv>=1.0.0
joblib>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytz>=2023.3
numba>=0.57.0
requests>=2.31.0
orjson>=3.9.0
fastparquet>=2023.4.0
pyarrow>=12.0.0
